+++ Packages/best.crux.avatar-sensing/CHANGELOG.md
@@ -7,10 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- A menu for footstep settings
+
 ### Fixed
 
 - Footstep sensors were off by default and had no controls to turn them on
-  - They are now on by default. Controls will be added later.
 
 ## [0.1.0] - 2025-10-17